% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.xml_document.R
\name{as_xml_document.sdmx_v2_1_structure_request}
\alias{as_xml_document.sdmx_v2_1_structure_request}
\title{Get data from a SDMX API}
\usage{
\method{as_xml_document}{sdmx_v2_1_structure_request}(x, ..., cache_dir = tempdir())
}
\arguments{
\item{x}{An httr2 request object}

\item{...}{reserved for future use}

\item{cache_dir}{The directory to cache the request in, set to \code{NULL} to disable caching}
}
\value{
a \code{xml2::xml_document} object
}
\description{
Retrieve data from an sdmx api and return it as an xml_document
}
\examples{
if (interactive()){
  dfs <- sdmx_v2_1_get_dataflows()
  flowRef <- dfs$ref[1]

  d <-
    sdmx_v2_1_structure_request(flowRef = flowRef) |>
    sdmx_v2_1_as_xml()

  d |>
    xml2::xml_find_all(".//generic:ObsValue") |>
    xml2::xml_text() |>
    head()

  path <- "example/dfs.json"
  d <-
    sdmx_v2_1_structure_request(
      detail = "referencestubs",
      # references = "all",
      format = "json"
    ) |>
    httr2::req_headers("Accept-Language" = "nl") |>
    httr2::req_perform(path=path)

}
}
