% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_request_from_url.R
\name{data_request_from_url}
\alias{data_request_from_url}
\title{Create a data request from a URL}
\usage{
data_request_from_url(url, ..., verbose = FALSE)
}
\arguments{
\item{url}{A URL that contains the parameters for the data request, including flowRef.}

\item{...}{Additional parameters to be passed to the data request.}

\item{verbose}{If \code{TRUE}, print information about the request.}
}
\value{
An object of class \code{\link[=sdmx_v2_1_data_request]{sdmx_v2_1_data_request()}} that can be used to retrieve data.
}
\description{
This function creates a data request from a URL that contains the necessary parameters for an SDMX data request.
Note that this is a raw request. Most users are better off using \code{\link[=get_observations_from_url]{get_observations_from_url()}} instead.
}
\examples{
# create a data request object from a URL
url <- "https://sdmx.oecd.org/public/rest/data/OECD.TAD.ARP,DSD_FISH_PROD@DF_FISH_AQUA,1.0/.A.._T.T?startPeriod=2010&dimensionAtObservation=AllDimensions"

# it does not execute the request, it just creates the request object
req <- data_request_from_url(url, verbose = TRUE)

# this is the raw csv output of the SDMX data request
# most users are better of using `get_observations_from_url()`
d <- req |> as.data.frame()
head(d)
}
