% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_to.R
\name{download_to}
\alias{download_to}
\title{Download the response of a request to a file}
\usage{
download_to(x, destfile, ...)
}
\arguments{
\item{x}{An httr2 request object}

\item{destfile}{A character string with the path to the file to save the response}

\item{...}{saved for future use}
}
\description{
Download the response of a request to a file
}
\details{
Download the response of a request to a file, but allows for different
format specifications e.g. csv, json, xml
}
