% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints with sdmx data}
\format{
A data.frame with columns:
\itemize{
\item \code{id} - the id of the endpoint
\item \code{name} - the name of agency of the endpoint
\item \code{url} - the url of the endpoint
\item \code{language} - the default language of the endpoint
}
}
\usage{
endpoints
}
\description{
A data.frame with several endpoints for sdmx data
}
\examples{
data("endpoints")
endpoints[, 1:3]

OECD <- get_endpoint("OECD")
OECD$verbose <- TRUE
list_oecd <- OECD |> list_dataflows(cache=TRUE)
}
\keyword{datasets}
