% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve data from a SDMX data API}
\usage{
get_data(
  endpoint = NULL,
  agencyID,
  id,
  version = "latest",
  ref = NULL,
  startPeriod = NULL,
  endPeriod = NULL,
  filter_on = list(),
  ...,
  dim_contents = c("label", "both", "id"),
  obs_value_numeric = TRUE,
  raw = FALSE,
  language = NULL,
  cache = TRUE,
  pivot = NULL,
  verbose = getOption("sdmxdata.verbose", FALSE)
)
}
\arguments{
\item{endpoint}{An endpoint, url or \code{\link[httr2:request]{httr2::request()}} object.}

\item{agencyID}{The agency ID}

\item{id}{The id of the dataflow}

\item{version}{The version of the dataflow, default "latest"}

\item{ref}{The dataflow reference can be used in stead of agencyID, id and version}

\item{startPeriod}{The start period for which the data should be returned, works only for dataflows with an explicit time dimension.}

\item{endPeriod}{The end period for which the data should be returned, works only for dataflows with an explicit time dimension.}

\item{filter_on}{A named list of filters to apply to the data, if not specified or \code{list()}, it is the default selection, set to \code{NULL} to select all.}

\item{...}{Additional parameters to pass to the request}

\item{dim_contents}{The contents of the dimension columns, either "label", "id" or "both"}

\item{obs_value_numeric}{Should the OBS_VALUE column be coerced to numeric? Default is \code{TRUE}}

\item{raw}{If \code{TRUE} return the raw data.frame from the SDMX, otherwise the data.frame is processed}

\item{language}{The language of the metadata}

\item{cache}{if \code{TRUE} cache the accompanying  meta data.}

\item{pivot}{\code{character} The name of the column to pivot the data on.
If \code{NULL} (default) the data is returned in a long format.}

\item{verbose}{If \code{TRUE} print information about the request}
}
\value{
a data.frame
}
\description{
Retrieve data from an SDMX data API.
The function retrieves data from the SDMX data API and returns it as a data.frame.
}
\details{
\code{get_data} is similar to \code{\link[=get_observations]{get_observations()}}, but its differs in two aspects:
\itemize{
\item it returns only data columns, no attribute columns
\item it can pivot the observations, i.e. it can return the data in a wide format, which
is useful feature when the data has multiple measures or when the data is to
presented as a time series.
}
}
