% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Connect to a SDMX endpoint}
\usage{
get_endpoint(id, language = NULL, singleton = TRUE, verbose = FALSE)
}
\arguments{
\item{id}{character, id of one of the available \code{endpoints}.}

\item{language}{character, the language to use for the text used in the response.}

\item{singleton}{logical, if \code{TRUE} return the same object if it already exists}

\item{verbose}{logical, if \code{TRUE} print information about the dataflows.}
}
\value{
a SDMXEndpoint object
}
\description{
Connect to an SDMX endpoint
}
\examples{
data("endpoints")
endpoints[, 1:3]

OECD <- get_endpoint("OECD")
OECD$verbose <- TRUE
list_oecd <- OECD |> list_dataflows(cache=TRUE)
}
\seealso{
Other SDMX endpoints: 
\code{\link{SDMXEndpoint}},
\code{\link{sdmx_endpoint}()}
}
\concept{SDMX endpoints}
