% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_by_url.R
\name{get_observations_by_url}
\alias{get_observations_by_url}
\title{Get observations using data-explorer url}
\usage{
get_observations_by_url(url, return_query = FALSE, verbose = FALSE)
}
\arguments{
\item{url}{character, the REST 2.1 url retrieved from an external source}

\item{return_query}{logical, if \code{TRUE} return the query object instead of the observation}

\item{verbose}{logical, if \code{TRUE} print information about the query}
}
\value{
when \code{return_query = FALSE} a data.frame with the observations, i.e.
same output as \code{\link[=get_observations]{get_observations()}}. Otherwise a list with the query object
and the parameters
}
\description{
Retrieve observations using the url from the data-explorer. The url given
will be parsed and translated into a \code{\link[=get_observations]{get_observations()}} call, to get you started.
The call will be printed to the console or can be retrieved by settings \code{return_query = TRUE}.
}
\examples{
# retrieve an external url pointing to sdmx data

url <- "https://sdmx-api.beta.cbs.nl/rest/data/NL1.CNVT,DF_85929NED,1.0/..2022JJ00+2023JJ00?dimensionAtObservation=AllDimensions"

obs <- get_observations_by_url(url)
head(obs)


query <- get_observations_by_url(url, return_query = TRUE)
query
}
