% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_from_url.R
\name{get_observations_from_url}
\alias{get_observations_from_url}
\title{Get observations using data-explorer url}
\usage{
get_observations_from_url(url, return_query = FALSE, verbose = FALSE)
}
\arguments{
\item{url}{character, the REST 2.1 url retrieved from an external source}

\item{return_query}{logical, if \code{TRUE} return the query object instead of the observation}

\item{verbose}{logical, if \code{TRUE} print information about the query}
}
\value{
when \code{return_query = FALSE} a data.frame with the observations, i.e.
same output as \code{\link[=get_observations]{get_observations()}}. Otherwise a list with the query object
and the parameters
}
\description{
Retrieve observations using the url from the data-explorer. The url given
will be parsed and translated into a \code{\link[=get_observations]{get_observations()}} call, to get you started.
The call will be printed to the console or can be retrieved by settings \code{return_query = TRUE}.
}
\examples{
# retrieve an external url pointing to sdmx data

url <-
  "https://sdmx.oecd.org/public/rest/data/OECD.TAD.ARP,DSD_FISH_PROD@DF_FISH_AQUA,1.0/.A.._T.T?startPeriod=2010&dimensionAtObservation=AllDimensions"

obs <- get_observations_from_url(url)
head(obs)

query <- get_observations_from_url(url, return_query = TRUE)
query
}
\seealso{
Other retrieve data: 
\code{\link{get_data}()},
\code{\link{get_observations}()}
}
\concept{retrieve data}
