% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{get_provider}
\alias{get_provider}
\title{Connect to a SDMX provider}
\usage{
get_provider(provider, language = NULL, verbose = FALSE)
}
\arguments{
\item{provider}{character, one of the available providers}

\item{language}{character, the language to use for the text used in the response.}

\item{verbose}{logical, if \code{TRUE} print information about the dataflows.}
}
\value{
a SDMXProvider object\
}
\description{
Connect to an SDMX provider
}
