% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_categoryschemes.R
\name{list_categoryschemes}
\alias{list_categoryschemes}
\title{Get sdmx category schemes}
\usage{
list_categoryschemes(
  req = NULL,
  agencyID = NULL,
  ...,
  language = "nl",
  cache_dir = tempdir(),
  raw = FALSE,
  verbose = getOption("sdmxdata.verbose", FALSE)
)
}
\arguments{
\item{req}{A character string or endpoint for a given endpoint.}

\item{agencyID}{A character string from a given agencyID.}

\item{...}{saved for future use.}

\item{language}{The language to use for the text used in the response.}

\item{cache_dir}{The directory to cache the request in, set to \code{NULL} to disable caching.}

\item{raw}{If \code{TRUE} return the raw data from the SDMX, otherwise the data is processed.}

\item{verbose}{if \code{TRUE} print information about the dataflows.}
}
\value{
a data.frame with available dataflows
}
\description{
Get sdmx categoryschemes from a given endpoint.
}
\examples{
dfs <- list_dataflows()

head(dfs)
}
