% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{providers}
\alias{providers}
\title{providers of sdmx data}
\format{
An object of class \code{data.frame} with 9 rows and 5 columns.
}
\usage{
providers
}
\description{
providers of sdmx data
}
\keyword{datasets}
