% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_endpoint.R
\name{sdmx_endpoint}
\alias{sdmx_endpoint}
\title{Returns the base request object}
\usage{
sdmx_endpoint(x, ...)
}
\arguments{
\item{x}{An endpoint or an httr2 request object}

\item{...}{saved for future use}
}
\value{
a \code{\link[httr2:request]{httr2::request()}} object
}
\description{
Returns the base request object
}
