% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_get_data.R
\name{sdmx_get_data}
\alias{sdmx_get_data}
\title{Get data from a SDMX API}
\usage{
sdmx_get_data(
  req,
  format = c("csv", "json"),
  labels = c("both", "id"),
  ...,
  as.data.table = FALSE
)
}
\arguments{
\item{req}{An httr2 request object}

\item{format}{The format of the data. Either "csv" or "json"}

\item{labels}{The labels to include in the data. Either "both" or "id"}

\item{...}{reserved for future use}

\item{as.data.table}{if TRUE return a \code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\value{
a data.frame or data.table depending on the value of \code{as.data.table}
}
\description{
Retrieve data from an sdmx api and return it as a data.frame
}
\examples{
if (interactive()){
  dfs <- sdmx_get_dataflows_v2_1()
  flowRef <- dfs$flowRefs[1]

  d <-
    sdmx_req_data_v2_1(flowRef = flowRef) |>
    sdmx_get_data()

  head(d)


  # json
  d <-
    sdmx_req_data_v2_1(flowRef = flowRef) |>
    sdmx_get_data(format = "json")

}
}
