% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_get_dataflows_v2_1.R
\name{sdmx_get_dataflows_v2_1}
\alias{sdmx_get_dataflows_v2_1}
\title{Get sdmx dataflows}
\usage{
sdmx_get_dataflows_v2_1(req = NULL, agencyID = NULL)
}
\arguments{
\item{req}{A character string from a given endpoint}

\item{agencyID}{A character string from a given agencyID}
}
\value{
a data.frame with available dataflows
}
\description{
Get sdmx dataflows from a given endpoint
}
