% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_v2_1_as_data_frame.R
\name{sdmx_v2_1_as_data_frame}
\alias{sdmx_v2_1_as_data_frame}
\title{Get data from a SDMX API}
\usage{
sdmx_v2_1_as_data_frame(
  req,
  format = c("csv", "json", "xml"),
  labels = c("both", "id"),
  ...,
  as.data.table = FALSE
)
}
\arguments{
\item{req}{An httr2 request object}

\item{format}{The format of the data. Either "csv" or "json"}

\item{labels}{The labels to include in the data. Either "both" or "id"}

\item{...}{reserved for future use}

\item{as.data.table}{if TRUE return a \code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\value{
a data.frame or data.table depending on the value of \code{as.data.table}
}
\description{
Retrieve data from an sdmx api and return it as a data.frame
}
\examples{
if (interactive()){
  dfs <- sdmx_v2_1_get_dataflows()
  flowRef <- dfs$ref[1]

  d <-
    sdmx_v2_1_data_request(flowRef = flowRef) |>
    sdmx_v2_1_as_data_frame()

  head(d)


  # json
  d <-
    sdmx_v2_1_data_request(flowRef = flowRef) |>
    sdmx_v2_1_as_data_frame(format = "json")

}
}
