% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_v2_1_as_xml.R
\name{sdmx_v2_1_as_xml}
\alias{sdmx_v2_1_as_xml}
\title{Get data from a SDMX API}
\usage{
sdmx_v2_1_as_xml(req, ...)
}
\arguments{
\item{req}{An httr2 request object}

\item{...}{reserved for future use}
}
\value{
a \code{xml2::xml_document} object
}
\description{
Retrieve data from an sdmx api and return it as an xml_document
}
\examples{
if (interactive()){
  dfs <- sdmx_v2_1_get_dataflows()
  flowRef <- dfs$ref[1]

  d <-
    sdmx_v2_1_data_request(flowRef = flowRef) |>
    sdmx_v2_1_as_xml()

  d |>
    xml2::xml_find_all(".//generic:ObsValue") |>
    xml2::xml_text() |>
    head()
}
}
