% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_v2_1_data_request.R
\name{sdmx_v2_1_data_request}
\alias{sdmx_v2_1_data_request}
\title{Request SDMX data}
\usage{
sdmx_v2_1_data_request(
  req = NULL,
  resource = c("data", "metadata"),
  flowRef = NULL,
  key = NULL,
  providerRef = NULL,
  ...,
  startPeriod = NULL,
  endPeriod = NULL,
  updatedAfter = NULL,
  firstNObservations = NULL,
  lastNObservations = NULL,
  dimensionAtObservation = NULL,
  detail = c("full", "dataonly", "serieskeysonly", "nodata"),
  includeHistory = NULL,
  labels = c("both", "id")
)
}
\arguments{
\item{req}{An endpoint or an httr2 request object}

\item{resource}{The resource to request. Either "data" or "metadata"}

\item{flowRef}{The flow reference to request, see details}

\item{key}{The key to request, see details}

\item{providerRef}{The provider reference to request, see details}

\item{...}{Additional parameters to pass to the request}

\item{startPeriod}{The start period for which the data should be returned}

\item{endPeriod}{The end period for which the data should be returned}

\item{updatedAfter}{Only return data that has been updated after this date}

\item{firstNObservations}{Only return the first n observations}

\item{lastNObservations}{Only return the last n observations}

\item{dimensionAtObservation}{The dimensions of the observations should be returned}

\item{detail}{The detail of the data to return. Either "full", "dataonly", "serieskeysonly" or "nodata"}

\item{includeHistory}{Include the different versions of the data}

\item{labels}{Include the labels in the data, only valid when format is "csv"}
}
\value{
a modified \code{\link[httr2:request]{httr2::request()}} object
}
\description{
\code{sdmx_v2_1_data_request} is a wrapper around the smdx rest api v2.1 and is used to retrieve
data from the api.
}
