% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_v2_1_get_dataflows.R
\name{sdmx_v2_1_get_dataflows}
\alias{sdmx_v2_1_get_dataflows}
\title{Get sdmx dataflows}
\usage{
sdmx_v2_1_get_dataflows(req = NULL, agencyID = NULL)
}
\arguments{
\item{req}{A character string from a given endpoint}

\item{agencyID}{A character string from a given agencyID}
}
\value{
a data.frame with available dataflows
}
\description{
Get sdmx dataflows from a given endpoint
}
