% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_v2_1_structure_request.R
\name{sdmx_v2_1_structure_request}
\alias{sdmx_v2_1_structure_request}
\title{Retrieve sdmx structure}
\usage{
sdmx_v2_1_structure_request(
  req = NULL,
  resource = "dataflow",
  agencyID = NULL,
  resourceID = NULL,
  version = NULL,
  itemID = NULL,
  format = c("json", "xml"),
  ...,
  detail = c("full", "allstubs", "referencestubs", "allcompletestubs",
    "referencecompletestubs", "referencepartial"),
  references = c("none", "parents", "parentsandsiblings", "children", "descendants",
    "all")
)
}
\arguments{
\item{req}{An endpoint or an httr2 request object}

\item{resource}{The resource to request. One of "datastructure", "metadatastructure", "categoryscheme", "conceptscheme", "codelist", "hierarchicalcodelist", "organisationscheme", "agencyscheme", "dataproviderscheme", "dataconsumerscheme", "organisationunitscheme", "dataflow", "metadataflow", "reportingtaxonomy", "provisionagreement", "structureset", "process", "categorisation", "contentconstraint", "attachmentconstraint", "actualconstraint", "allowedconstraint", "structure", "transformationscheme", "rulesetscheme", "userdefinedoperatorscheme", "customtypescheme", "namepersonalisationscheme", "vtlmappingscheme"}

\item{agencyID}{The agency ID to request}

\item{resourceID}{The resource ID to request}

\item{version}{The version to request}

\item{itemID}{The item ID to request}

\item{format}{The format to request. Either "json" or "xml"}

\item{...}{saved for future use}

\item{detail}{The detail of the data to return. Either "full", "allstubs", "referencestubs", "allcompletestubs", "referencecompletestubs", "referencepartial"}

\item{references}{The references to return. Either "none", "parents", "parentsandsiblings", "children", "descendants", "all"}
}
\value{
a modified \code{\link[httr2:request]{httr2::request()}} object
}
\description{
Wrapper for the SDMX REST v2.1 interface for structured information
}
