% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXEndpoint.R
\name{SDMXEndpoint}
\alias{SDMXEndpoint}
\title{Base class for SDMX endpoints}
\description{
Base class for SDMX endpoints
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{req}}{\code{\link[httr2:request]{httr2::request()}} object, can be adjusted to add headers or proxy settings}

\item{\code{version}}{character, the SDMX version of the endpoint}

\item{\code{id}}{character, the id of the endpoint}

\item{\code{name}}{character, the name of the endpoint}

\item{\code{url}}{character, the url of the endpoint}

\item{\code{verbose}}{logical, if \code{TRUE} print information the connection}

\item{\code{cache_dir}}{character, the directory to cache the metadata in}

\item{\code{language}}{character, the language to use for the text used in the response}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SDMXEndpoint-new}{\code{SDMXEndpoint$new()}}
\item \href{#method-SDMXEndpoint-list_dataflows}{\code{SDMXEndpoint$list_dataflows()}}
\item \href{#method-SDMXEndpoint-get_dataflow_structure}{\code{SDMXEndpoint$get_dataflow_structure()}}
\item \href{#method-SDMXEndpoint-get_observations}{\code{SDMXEndpoint$get_observations()}}
\item \href{#method-SDMXEndpoint-get_data}{\code{SDMXEndpoint$get_data()}}
\item \href{#method-SDMXEndpoint-list_agencies}{\code{SDMXEndpoint$list_agencies()}}
\item \href{#method-SDMXEndpoint-clone}{\code{SDMXEndpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-new}{}}}
\subsection{Method \code{new()}}{
create a new SDMXEndpoint object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$new(
  url,
  id = gsub("\\\\W+", "_", url),
  name = url,
  language = NULL,
  cache_dir = tempdir(),
  verbose = getOption("sdmxdata.verbose", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character, the url of the endpoint}

\item{\code{id}}{character, the id of the endpoint}

\item{\code{name}}{character, the name of the endpoint}

\item{\code{language}}{character, the language to use for the text used in the response}

\item{\code{cache_dir}}{character, the directory to cache the metadata in}

\item{\code{verbose}}{logical, if \code{TRUE} print information on the connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-list_dataflows"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-list_dataflows}{}}}
\subsection{Method \code{list_dataflows()}}{
list the dataflows of an endpoint
Same as \code{\link[=list_dataflows]{list_dataflows()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$list_dataflows(
  agencyID = NULL,
  ...,
  raw = FALSE,
  language = self$language,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agencyID}}{character, the agencyID to filter on}

\item{\code{...}}{additional parameters to pass to the \code{\link[=list_dataflows]{list_dataflows()}} function}

\item{\code{raw}}{logical, if \code{TRUE} return the raw SDMX json data.}

\item{\code{language}}{character, the language to use for the text used in the response}

\item{\code{verbose}}{logical, if \code{TRUE} print information about the dataflows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{\link[=data.frame]{data.frame()}} with the dataflows,
get the dataflow structure of an endpoint

get the dataflow structure of an endpoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-get_dataflow_structure"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-get_dataflow_structure}{}}}
\subsection{Method \code{get_dataflow_structure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$get_dataflow_structure(
  ref = NULL,
  id = NULL,
  agencyID = NULL,
  version = "latest",
  cache = (!is.null(self$cache_dir)),
  language = self$language
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-get_observations"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-get_observations}{}}}
\subsection{Method \code{get_observations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$get_observations(
  agencyID = NULL,
  id = NULL,
  version = "latest",
  ref = NULL,
  filter_on = list(),
  ...,
  cache = (!is.null(self$cache_dir)),
  language = self$language
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$get_data(
  agencyID = NULL,
  id = NULL,
  ref = NULL,
  filter_on = list(),
  pivot = NULL,
  ...,
  cache = (!is.null(self$cache_dir)),
  language = self$language,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-list_agencies"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-list_agencies}{}}}
\subsection{Method \code{list_agencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$list_agencies(
  ...,
  cache = (!is.null(self$cache_dir)),
  raw = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
