% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_endpoint.R
\name{search_endpoint}
\alias{search_endpoint}
\title{Get the search endpoint}
\usage{
search_endpoint(x, ...)
}
\arguments{
\item{x}{An endpoint or an httr2 request object}

\item{...}{saved for future use}
}
\value{
a \code{\link[httr2:request]{httr2::request()}} object
}
\description{
Get the search endpoint
}
