% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_sdmx_v2_1_data_request.R
\name{as.data.frame.sdmx_v2_1_data_request}
\alias{as.data.frame.sdmx_v2_1_data_request}
\title{Get data from a SDMX API}
\usage{
\method{as.data.frame}{sdmx_v2_1_data_request}(
  x,
  row.names = NULL,
  optional = FALSE,
  format = c("csv", "json", "xml"),
  labels = c("both", "id"),
  ...,
  language = "nl",
  as.data.table = FALSE
)
}
\arguments{
\item{format}{The format of the data. Either "csv" or "json"}

\item{labels}{The labels to include in the data. Either "both" or "id"}

\item{...}{reserved for future use}

\item{language}{The language to use for the text used in the response}

\item{as.data.table}{if TRUE return a \code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{req}{An httr2 request object}
}
\value{
a data.frame or data.table depending on the value of \code{as.data.table}
}
\description{
Retrieve data from an sdmx api and return it as a data.frame
}
\examples{
if (interactive()){
  dfs <- sdmx_v2_1_get_dataflows()
  flowRef <- dfs$ref[2]

  system.time({
    d <-
      sdmx_v2_1_data_request(flowRef = flowRef)

    sdmx_v2_1_as_data_frame()
  })


  # json
  d <-
    sdmx_v2_1_data_request(flowRef = flowRef) |>
    sdmx_v2_1_as_data_frame(format = "json")

}
}
