% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataflow_structure.R
\name{get_dataflow_structure}
\alias{get_dataflow_structure}
\title{Get information about a dataflow}
\usage{
get_dataflow_structure(
  req = NULL,
  ref,
  agencyID = getOption("cbsopendata.agencyID", NULL),
  id,
  version = "latest",
  language = getOption("cbsopendata.language", "en"),
  cache_dir = tempdir(),
  verbose = getOption("cbsopendata.verbose", FALSE)
)
}
\arguments{
\item{req}{A character string with the request to the dataflow}

\item{ref}{A string with the flow reference, in the form of "agencyID:id(version)". Overrules
the agencyID, id and version arguments.}

\item{agencyID}{The agency ID of the dataflow}

\item{id}{The id of the dataflow}

\item{version}{The version of the dataflow, defaults to "latest"}

\item{language}{The language of the metadata}

\item{cache_dir}{The directory to cache the data in, set to \code{NULL} to disable caching.}

\item{verbose}{print some information on the console}
}
\value{
a list with the dataflow information
}
\description{
Get information and structure  about a dataflow. The dataflow can be identified
using \code{agencyID}, \code{id} and \code{version} or by using the \code{ref} argument.
}
\details{
For most use cases the \code{agencyID} and \code{id} are to be preferred, as the \code{ref} argument
includes a specific version number of the dataflow. Not specifying a version number
will default to the latest version. If it is desirable to pin the reference of
the dataflow to a specific version, the \code{ref} argument can be used. The \code{ref}
argument can also be found in \code{\link[=list_dataflows]{list_dataflows()}}.
}
