% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataflow_structure.R
\name{get_dataflow_structure}
\alias{get_dataflow_structure}
\title{Get information about a dataflow}
\usage{
get_dataflow_structure(
  req = NULL,
  flowRef,
  id,
  version = "latest",
  agencyID = getOption("sdmxdata.agencyID", "NL1"),
  language = getOption("sdmxdata.language", "nl"),
  cache_dir = tempdir(),
  verbose = getOption("sdmxdata.verbose", FALSE)
)
}
\arguments{
\item{req}{A character string with the request to the dataflow}

\item{flowRef}{A string with the flow reference, in the form of "agencyID,id,version"}

\item{id}{The id of the dataflow}

\item{version}{The version of the dataflow, defaults to "latest"}

\item{agencyID}{The agency ID of the dataflow, defaults to "NL1"}

\item{language}{The language of the metadata, defaults to "nl"}

\item{cache_dir}{The directory to cache the data in, set to \code{NULL} to disable caching.}

\item{verbose}{print some information on the console}
}
\value{
a list with the dataflow information
}
\description{
Get infomration about a dataflow
}
