% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_agencies.R
\name{list_agencies}
\alias{list_agencies}
\title{list the agencies of an endpoint}
\usage{
list_agencies(endpoint = NULL, language = NULL, raw = FALSE, cache = TRUE)
}
\arguments{
\item{endpoint}{an endpoint, url or an \code{\link[httr2:request]{httr2::request()}} object}

\item{language}{the language to return the data in}

\item{raw}{if \code{TRUE} return the raw data}

\item{cache_dir}{the directory to cache the data in, set to \code{NULL} to disable caching}

\item{verbose}{if \code{TRUE} print information about the caching.}
}
\description{
list the agencies of an endpoint
}
