% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXEndpoint.R
\name{SDMXEndpoint}
\alias{SDMXEndpoint}
\title{Base class for SDMX endpoints}
\description{
Base class for SDMX endpoints
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{req}}{\code{\link[httr2:request]{httr2::request()}} object, can be adjusted to add headers or proxy settings}

\item{\code{version}}{character, the SDMX version of the endpoint}

\item{\code{id}}{character, the id of the endpoint}

\item{\code{name}}{character, the name of the endpoint}

\item{\code{url}}{character, the url of the endpoint}

\item{\code{verbose}}{logical, if \code{TRUE} print information the connection}

\item{\code{cache_dir}}{character, the directory to cache the metadata in}

\item{\code{language}}{character, the language to use for the text used in the response}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SDMXEndpoint-new}{\code{SDMXEndpoint$new()}}
\item \href{#method-SDMXEndpoint-clone}{\code{SDMXEndpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-new}{}}}
\subsection{Method \code{new()}}{
create a new SDMXEndpoint object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$new(
  url,
  id = gsub("\\\\W+", "_", url),
  name = url,
  language = NULL,
  cache_dir = file.path(tempdir(), "sdmxdata", id),
  verbose = getOption("sdmxdata.verbose", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character, the url of the endpoint}

\item{\code{id}}{character, the id of the endpoint}

\item{\code{name}}{character, the name of the endpoint}

\item{\code{language}}{character, the language to use for the text used in the response}

\item{\code{cache_dir}}{character, the directory to cache the metadata in}

\item{\code{verbose}}{logical, if \code{TRUE} print information on the connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDMXEndpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SDMXEndpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDMXEndpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
